+++ lib/chukinas_web/live/dreadnought_liveview.ex
@@ -1,0 +1,74 @@
+defmodule ChukinasWeb.DreadnoughtLiveViewHelpers do
+  @moduledoc"""
+  Provides common functions to DreadnoughtLive and DreadnoughtPlayLive
+  """
+
+  alias ChukinasWeb.Router.Helpers, as: Routes
+  alias Chukinas.Sessions
+
+  defmacro __using__(_opts) do
+    quote do
+
+      use ChukinasWeb, :live_view
+      alias Chukinas.Dreadnought.Mission
+
+      # *** *******************************
+      # *** CALLBACKS
+
+      @impl true
+      def handle_info({:push_redirect, path}, socket) do
+        socket
+        |> push_redirect(to: path)
+        |> noreply
+      end
+
+      @impl true
+      def handle_info({:update_mission, mission}, socket) do
+        if mission_in_progress?(mission) do
+          path = Routes.dreadnought_main_path(socket, :play)
+          socket
+          |> Phoenix.LiveView.push_redirect(to: path)
+          |> noreply
+        else
+          socket
+          |> assign(mission: mission)
+          |> noreply
+        end
+      end
+
+      # *** *******************************
+      # *** HELPERS
+
+      @spec assign_uuid_and_mission(Phoenix.LiveView.Socket.t, map) :: Phoenix.LiveView.Socket.t
+      def assign_uuid_and_mission(socket, session) do
+        uuid = Map.fetch!(session, "uuid")
+        if socket.connected? do
+          Sessions.register_uuid(uuid)
+          socket
+          |> assign(uuid: uuid)
+          |> assign(mission: Sessions.get_mission_from_player_uuid(uuid))
+        else
+          socket
+          |> assign(uuid: uuid)
+          |> assign(mission: nil)
+        end
+      end
+
+      def mission_in_progress?(socket) do
+        with %Chukinas.Dreadnought.Mission{} = mission <- mission(socket),
+             true <- Chukinas.Dreadnought.Mission.in_progress?(mission) do
+          true
+        else
+          _ -> false
+        end
+      end
+
+      # TODO this is ugly
+      def mission(nil), do: nil
+      def mission(%Chukinas.Dreadnought.Mission{} = value), do: value
+      def mission(socket), do: socket.assigns[:mission]
+
+    end
+  end
+
+end
